Changes To Migrate To Cosmos Db - Closes Issue 19
This is quite a big commit, which probably should have been split into smaller, but I was in a hurry. The idea is that this way there is a modular way to bring in either a sql database or a no sql database and have the codebase with minor changes in configuration work flawlessly(hopefully). For this reason shared models library had to go and in its place multiple models were created in each model to replace it. Additionally a REST API was added between the gateway and the data library for even more modularity and for preparation to a transition to microservice architecture. For NoSql combatability the standard id of the post entity has also been replaced with GUID.