Changed Email Library To A Microservice - Closes Issue 27
I changed the email library to a microservice by adding a rest API as its gateway project. The email library is a bit more sophisticated and it also saves the emails to the database for possible later inspection. As with the dataaccess microservice this microservice supports as its dbms both sql(sqlserver) and nosql(cosmosdb), but in production for costs reasons we will go with cosmos.