Added Integration Tests For Email Rest API - Closes Issue 33
In this update a new test project has been added and in that 7 integration tests have been also added, which test the 4 endpoints of the Email Rest API. These test will make sure concistency in future changes that might be added to the library. Some minor configuration updates have also been added to the project as a whole and some solution filters for easier management of the different microservices that are contained in the project.