Added Integration Tests For The Data Microservice - Closes Issue 35
In this commit I have added integration tests that test all the end points of the Data Microservice REST API similarly to the integration tests I added to the Email Microservice REST API in the previous commit. I also added a GetPost endpoint to the REST API that was missing and refactor a bit the status codes it returned in some of its endpoints.