Added Unit Tests For Data Microservice - Closes Issue 42
Added units that test all the endpoints of the Rest API to prevent code regressions in further commits. Also improved the E2E, which can now run from the test runner the same way the integration and unit tests run. The only possible issue with the test runner is that the E2E should not run at the same time as the integration tests, since running them like that could lead to conflicts.